Add roster CI workbook, extended GymAware ETL, and BI extract layer
- Commit data/roster/roster_new.xlsx and coach handoff docs; sync roster to Supabase before scheduled ETL.

- GymAware /reps, /athletes, /bests export and upload with BI extract tables.

- DDL for catapult_stats_bi_extract, whoop_bi_extract, gymaware_extended, identity BIGINT fix, legacy dashboard cleanup.

- Docs, workflow, and .env.example aligned with online GitHub Actions runs.

+++ schema/athlete_identity.sql
@@ -6,7 +6,7 @@ CREATE TABLE IF NOT EXISTS public.athlete_identity (
     internal_key TEXT NOT NULL,
     display_name TEXT,
     catapult_athlete_id TEXT,
-    gymaware_athlete_reference INTEGER,
+    gymaware_athlete_reference BIGINT,
     vald_profile_id TEXT,
     whoop_user_id TEXT,
     teamworks_athlete_id TEXT,
